Rewrite copy commands (#1873)
* Allow `SyncCommandBufferBuilder` commands to specify the access range

* Allow `SyncCommandBufferBuilder` commands to specify the access range

* Better handling of image initialization

* Small fix

* Rewrite copy commands

* Better handling of maintenance1

* Small leftovers

* Remove conflict_key

+++ vulkano/src/swapchain/swapchain.rs
@@ -688,7 +688,7 @@ impl<W> Swapchain<W> {
         self.images.get(offset).map(|i| ImageInner {
             image: &i.image,
             first_layer: 0,
-            num_layers: self.image_array_layers as usize,
+            num_layers: self.image_array_layers,
             first_mipmap_level: 0,
             num_mipmap_levels: 1,
         })
